fix: fixed bugs inside transaction.c

+++ src/utils/sys_utils.h
@@ -1,7 +1,7 @@
 #ifndef MINIMALIST_BLOCK_CHAIN_SYSTEM_SRC_UTILS_SYS_UTILS_H
 #define MINIMALIST_BLOCK_CHAIN_SYSTEM_SRC_UTILS_SYS_UTILS_H
 
-int get_timestamp();
+unsigned long get_timestamp();
 char *get_str_timestamp(char *, unsigned int);
 
 #endif
